massive -game- -phaser- code clean-up and refactoring for better maintanability
